==> analysis/03-summarize_vert_meas.Rmd <==
---
title: "Summarize vertebral measurements"
output:
  html_document:
    df_print: paged
---

```{r setup}
library(tidyverse)
library(emmeans)
library(ggbeeswarm)
library(patchwork)
library(here)
```

# Load data

## Vertebral measurements

```{r}
vertdata <- read_csv(here('output', "MasterVert_Measurements_Matched.csv")) %>%
  separate(MatchSpecies, into=c("MatchGenus", "MatchSpecies"), sep="_") %>%
  relocate(MatchGenus, MatchSpecies, .after=Species) %>%
  rename(alphaPos = alpha_Pos,
         alphaAnt = alpha_Ant,
         DPos = D_Pos,
         DAnt = D_Ant,
         BodyShape = `Body Shape`,
         dBW = d_BW,
         DAntBW = D_Ant_BW,
         DPosBW = D_Pos_BW,
         fineness = `SL/Max_BW`)

head(vertdata)
```

```{r}
vertdata |> 
  filter(Habitat_Friedman != Habitat) |> 
  distinct(Species, .keep_all = TRUE) |> 
  select(Species, starts_with('Habitat')) |> 
  arrange(Habitat_Friedman)
```

```{r}
vertdata |>
  unite(MatchSpecies, MatchGenus, MatchSpecies) |> 
  filter(Indiv == 1) |> 
  group_by(MatchSpecies, Pos) |> 
  filter(n() > 1)

```
```{r}
vertdata <-
  vertdata |> 
  filter(Indiv == 1 & Species != 'Xiphister_atropurpureus')
```


Compute the correct CBL
```{r}
vertdata <-
  vertdata %>%
  rename(CBL_old_raw = CBL_raw,
         CBL_old = CBL) %>%
  mutate(CBL_raw = (abs(Pt1x - Pt3x) + abs(Pt2x - Pt4x))/2,
         CBL = CBL_raw / SL)
```

```{r}
ggplot(vertdata, aes(x = CBL_old, y = CBL)) +
  geom_point()
```
The values are similar but not identical.

# Basic plots

```{r}
vertdata <-
  vertdata %>%
  mutate(d_normD = d / ((DAnt + DPos)/2),
         d_normCBL = d / CBL,
         Iratio = 1 - d^4/((DAnt + DPos)/2)^4) %>%
  mutate(Pos = Pos/100,
         Pos = factor(Pos))
```

Calculate the mean of each variable at each body position.
```{r}
vertdata_bypos <-
  vertdata %>%
  group_by(Pos) %>%
  dplyr::summarize(across(c(d, CBL, alphaAnt, alphaPos, DAnt, DPos, dBW, DPosBW, DAntBW, 
                            d_normCBL, d_normD, Iratio), 
                   list(mn = ~ mean(.x, na.rm = TRUE),
                        med = ~ median(.x, na.rm = TRUE),
                        iqr = ~ IQR(.x, na.rm = TRUE),
                        sd = ~ sd(.x, na.rm = TRUE))))
```

```{r}
vertdata_bypos %>%
  pivot_longer(!Pos, names_to=c("var", ".value"),
               names_pattern = "(.*)_(.*)") %>%
  arrange(var, Pos) %>%
  ggplot(aes(x = Pos, y = med, group = 1)) +
  geom_ribbon(aes(ymin = med-iqr, ymax = med+iqr), alpha = 0.5) +
  geom_line() +
  geom_line(aes(y = mn), color="red") +
  facet_wrap(~ var, scales = "free")
```

```{r}
vertdata <-
  vertdata %>%
  group_by(Pos) %>%
  mutate(across(c(d, CBL, alphaAnt, alphaPos, DAnt, DPos, dBW, DAntBW, DPosBW, d_normD, d_normCBL, Iratio), 
                   list(ctr = ~.x - median(.x, na.rm = TRUE)))) %>%
  ungroup()
```

```{r}
vertdata %>%
  ggplot(aes(x = Pos, y = d_ctr, color=Habitat)) +
  geom_boxplot(aes(group = Pos)) +
  geom_beeswarm() +
  geom_line(data = filter(vertdata, d > 0.006 & Habitat == "pelagic"),
            aes(group=Species, linetype=Species)) +
  geom_label(data = filter(vertdata, d > 0.006 & Pos == 0.9 & Habitat == "pelagic"),
             aes(label = Species)) +
  facet_grid(. ~ Habitat) +
  labs(x = "Vertebra position (L)",
       y = "Foramen diameter (L)")
```

```{r}
vertdata %>%
  ggplot(aes(x = Pos, y = dBW_ctr, color=Habitat)) +
  geom_boxplot(aes(group = Pos)) +
  geom_beeswarm() +
  geom_line(data = filter(vertdata, dBW_ctr > 0.05 & Habitat == "pelagic"),
             aes(group=Species, linetype=Species)) +
  geom_label(data = filter(vertdata, dBW_ctr > 0.05 & Pos == 0.9 & Habitat == "pelagic"),
              aes(label = Species)) +
  facet_grid(. ~ Habitat) +
  labs(x = "Vertebra position (L)",
       y = "Foramen diameter (L)")
```

Let's exclude Elops and Mirorictus from the rest of the analysis for right now.
```{r}
vertdata2 <-
  vertdata %>%
  filter((Species != "Elops_saurus") &
           Species != "Mirorictus_taningi")
```

```{r}
vertdata2 %>%
  filter(is.na(d_normCBL))
```

```{r}
write_csv(vertdata2, here('output', "vertdata_centered.csv"))
```


```{r}
plot_position_habitat_distribution <- function(df, var) 
{
  var <- enquo(var)

  p1 <-
    df %>%
    filter(!is.na(!!var)) %>%
    ggplot(aes(x = Pos, y = !!var, color=Habitat, fill=Habitat, group=Habitat)) +
    stat_summary(fun.data = "mean_se", geom="ribbon", alpha=0.5) +
    stat_summary(fun = "mean", geom="line")
  
  p2 <-
    df %>%
    filter(!is.na(!!var)) %>%
    ggplot(aes(x = Habitat, y = !!var, color=Habitat)) +
    geom_violin() +
    geom_boxplot(width=0.3, alpha=0.5) +
    stat_summary(aes(group = 1), fun = "median", geom = "line")
  
  p1 + p2 + plot_layout(widths = c(3,1), guides = 'collect')  
}
```

```{r}
plot_position_habitat_distribution(vertdata2, d)
```
Here we're plotting the foramen diameter (normalized by body length) relative to position on the left, and the overall distributions relative to habitat on the right. The bottom row has the overall mean pattern relative to body length subtracted.

```{r}
plot_position_habitat_distribution(vertdata2, dBW)
```

```{r}
plot_position_habitat_distribution(vertdata2, CBL)
```

```{r}
plot_position_habitat_distribution(vertdata2, alphaAnt)
```

```{r}
plot_position_habitat_distribution(vertdata2, alphaPos)
```

```{r}
plot_position_habitat_distribution(vertdata2, DAnt)
```

# Compare summary statistics

First generate the summary statistics, summarizing across the body positions.

This gives us the mean, median, and max values for each of the measurements.
```{r}
vertdata_summary <-
  vertdata2 %>%
  dplyr::filter(Pos != 0.2 & Pos != 0.3) %>%
  group_by(Species, Indiv, Habitat, Water_Type, MatchSpecies, MatchGenus, 
           alltaxon, Order, Family) %>%
  dplyr::summarize(fineness = mean(fineness),
                   across(c(CBL, d, alphaAnt, alphaPos, DAnt, DPos,
                            dBW, DAntBW, DPosBW, d_normCBL, d_normD, Iratio),
                          list(med = ~ median(.x, na.rm = TRUE),
                               max = ~ max(.x, na.rm = TRUE),
                               mn = ~ mean(.x, na.rm = TRUE)))) %>%
  ungroup()
```

```{r}
head(vertdata_summary)
```

```{r}
vertdata_summary %>%
  select(ends_with("max") | Habitat | Species) %>%
  pivot_longer(contains("max"),
               names_to = "var", values_to = "value") %>%
  ggplot(aes(x = Habitat, y = value, color = Habitat)) +
  geom_violin() +
  geom_boxplot(width=0.3, alpha=0.5) +
  stat_summary(aes(group = 1), fun = "median", geom = "line") +
  facet_wrap(~ var, scales = "free")
```

```{r}
write_csv(vertdata_summary, here("output", "vertdata_summary.csv"))
```
