==> analysis/11-VertPGLS.Rmd <==
---
title: "R Notebook"
output:
  html_document:
    df_print: paged
---

```{r setup}
library(tidyverse)
library(ggbeeswarm)
library(phytools)
library(patchwork)
library(here)
library(nlme)
library(ape)
library(geiger)
library(ggtree)
library(emmeans)
library(car)
library(Hmisc)
```

```{r}
citation()
print(getRversion())
```

```{r}
citation("nlme")
packageVersion("nlme")
```
```{r}
citation("ape")
packageVersion("ape")
```

```{r}
citation("geiger")
packageVersion("geiger")
```

# Load data

```{r}
vertdata_sum <- read_csv(here("output/vertdata_summary_lm.csv"))
```

## Phylogeny

This is the whole Betancur-R tree.
```{r}
tree <- read.tree(here('data/12862_2017_958_MOESM2_ESM.tre'))
```

Get the names of species from the tree.
```{r}
allspecies <- tibble(tree$tip.label)
colnames(allspecies) <- c('FullName')
head(allspecies)
```

And split the names into family, genus, and species.
```{r}
allspecies <- 
  allspecies %>% separate(FullName, sep='_', into=c('Family', 'Genus', 'Species'), 
                          extra='drop', remove=FALSE)
```

Set up the tip number (just the row)
```{r}
allspecies$Tip <- seq_len(nrow(allspecies))
```

```{r}
vertdata <- left_join(vertdata_sum, allspecies, 
                                  by=c("MatchGenus"="Genus", "MatchSpecies"="Species")) %>%
  select(-(ends_with(".x") | ends_with(".y")))
vertdata
```

Drop species without a match
```{r}
vertdata %>%
  filter(is.na(Tip)) %>%
  distinct(Species, .keep_all=TRUE) %>%
  select(Species, MatchGenus, MatchSpecies, Tip, Habitat)
```

```{r}
vertdata <-
  vertdata %>%
  filter(!is.na(Tip))
```


```{r}
ourspecies <-
  vertdata %>%
  distinct(Species, .keep_all=TRUE)
```


```{r}
verttree <- keep.tip(tree, tip=as.vector(ourspecies$Tip))
```

```{r}
plotTree(verttree)
```


```{r}
vertdata_sp <- 
  vertdata %>%
  distinct(FullName, .keep_all = TRUE) %>%
  mutate(rowname = FullName) %>%
  column_to_rownames(var = "rowname")
```

```{r}
verttree <- keep.tip(tree, tip=as.vector(vertdata_sp$Tip))
```

```{r}
left_join(as_tibble(verttree),
          vertdata_sp %>%
            rownames_to_column("label") %>%
            select(label, Habitat)) %>%
  tidytree::as.treedata() %>%
  ggtree(layout = 'circular') + # geom_tiplab() +
  geom_tippoint(aes(color = Habitat))
```
Check if tree and data match
```{r}
length(verttree$tip.label)
nrow(vertdata_sp)
```

```{r}
name.check(verttree, vertdata_sp)
```

Merge the measurements and the tree
```{r}
verttree_data <-
  as_tibble(verttree) %>%
  left_join(vertdata_sp %>%
              rownames_to_column("label") %>%
              select(label, Habitat, Family,
                     ends_with("80"), ends_with("max"), ends_with("med"), ends_with("slope"), ends_with("quad")))
```

```{r}
vertdata_sp <-
  vertdata_sp %>%
  mutate(CBL_vtx = -CBL_slope / CBL_quad,
         alphaPos_vtx = -alphaPos_slope / alphaPos_quad,
         alphaAnt_vtx = -alphaAnt_slope / alphaAnt_quad,
         d_vtx = -d_slope / d_quad,
         DAnt_vtx = -DAnt_slope / DAnt_quad,
         DPos_vtx = -DPos_slope / DPos_quad)
```

Save out the tree data. There is something subtly different about saving the data out as a csv file and saving it as an RDS file. The base class of the tree is tbl_tree, but when we load it back in from a csv, despite having exactly the same data, it won't work with the tidytree functions. So we save it in an RDS file, which preserves the class.
```{r}
write_csv(vertdata_sp, here('output/vertdata_summary_lm_species.csv'))
saveRDS(verttree, here('output/vert_tree.rds'))
```

## PGLS Analysis for means, maxes, medians, and 80%

```{r}
vertdata_sp0 <-
  vertdata_sp %>%
  mutate(across(contains('slope') | contains('quad'), 
                ~ replace_na(., 0)))
```

```{r}
var <- c(#'CBL_80', 'd_80', 'alphaAnt_80', 'alphaPos_80', 'DAnt_80', 'DPos_80', 
         #'CBL_max', 'd_max', 'alphaAnt_max', 'alphaPos_max', 'DAnt_max', 'DPos_max',
         'CBL_mn', 'd_mn', 'alphaAnt_mn', 'alphaPos_mn', 'DAnt_mn', 'DPos_mn',
         #'CBL_med', 'd_med', 'alphaAnt_med', 'alphaPos_med', 'DAnt_med', 'DPos_med',
         'fineness', 
         #'CBL_slope', 'd_slope', 'alphaAnt_slope', 'alphaPos_slope', 'DAnt_slope', 'DPos_slope', 
         'CBL_vtx', 'd_vtx', 'alphaAnt_vtx', 'alphaPos_vtx', 'DAnt_vtx', 'DPos_vtx',
         'CBL_quad', 'd_quad', 'alphaAnt_quad', 'alphaPos_quad', 'DAnt_quad', 'DPos_quad')
         #'CBL_order', 'd_order', 'alphaAnt_order', 'alphaPos_order', 'DAnt_order', 'DPos_order')

modeltests <- tibble()

for (i in seq_along(var)) {
  print(var[i])
  fmla <- as.formula(paste(var[i], "Habitat", sep = " ~ "))
  mod <- gls(fmla, correlation = corBrownian(1, phy = verttree, form = ~FullName),
                      data = vertdata_sp0, method = "ML")
  
  ava <- broom::tidy(Anova(mod))
  ava$var = var[i]
  ava$model = list(mod)
  
  modeltests <- bind_rows(modeltests, ava)
}

modeltests %>%
  dplyr::select(var, statistic, p.value, everything()) %>%
  filter(p.value < 0.05 | str_detect(var, "dBW")) %>%
  arrange(var)
```

```{r}
compare_habitats <- function(model) {
  emm <- emmeans(model, ~Habitat)
  p <- as.data.frame(pairs(emm))
  es <- as.data.frame(eff_size(emm, sigma = sigma(model), edf = model$dims$N - model$dims$p))
  p$effect.size <- es$effect.size
  
  p %>%
    mutate(contrast = str_replace(contrast, "(\\w+) - (\\w+)", "\\1_\\2")) %>%
    select(contrast, p.value, effect.size) %>%
    rename(p = p.value, eff = effect.size) %>%
    pivot_wider(names_from = contrast, values_from = c(p, eff),
                names_glue = "{contrast}_{.value}")
}
```

```{r}
habitat_means <- function(model) {
  emm <- emmeans(model, ~Habitat)
  as.data.frame(emm) %>%
    select(Habitat, emmean, SE) %>%
    rename(mn = emmean,
           se = SE) %>%
    pivot_wider(names_from = Habitat, values_from = c(mn, se),
                names_glue = "{Habitat}_{.value}")
    
}
```

```{r}
modeltests <-
  modeltests %>%
  dplyr::select(var, statistic, p.value, everything()) %>%
  # filter(p.value < 0.05) %>%
  mutate(mc = purrr::map(model, compare_habitats)) %>%
  unnest(mc) %>%
  mutate(total_eff = abs(benthic_demersal_eff) + abs(benthic_pelagic_eff) + abs(demersal_pelagic_eff)) %>%
  mutate(hm = purrr::map(model, habitat_means)) %>%
  unnest(hm) %>%
  # filter(total_eff > 0.5) %>%
  relocate(p.value, total_eff, ends_with("_p"), .after = var) %>%
  arrange(desc(total_eff))
```

```{r}
modeltests %>%
  filter(abs(benthic_demersal_eff) >= 0.2 | abs(benthic_pelagic_eff) >= 0.2 | abs(demersal_pelagic_eff) >= 0.2)
```
```{r}
emmeans(modeltests$model[[2]], ~Habitat)
```

```{r}
sigvars <- pull(modeltests, var)
```

```{r}
write_csv(modeltests, here("output/modeltests.csv"))
saveRDS(modeltests, here("output/PGLSmodels.Rds"))
```

## Mean plots with stats

This sets up the data in long form, with all the variables stacked, and the variables ordered by descending total effect size.
```{r}
vertdata_long <-
  vertdata_sp0 %>%
  select(Habitat, Species, !!sigvars) %>%
  pivot_longer(cols = !Habitat & !Species, names_to = "var", values_to = "value") %>%
  left_join(
    modeltests %>%
      select(var, total_eff, p.value), 
    by = "var") %>%
  mutate(var = as.factor(var),
         var = fct_reorder(var, total_eff, .desc = TRUE))
  
```

Plot all the effects
```{r}
vertdata_long %>%
  filter(!is.na(var) & p.value < 0.05) %>%
  ggplot(aes(x = Habitat, y = value, color = Habitat)) +
  geom_quasirandom(width=0.3, alpha = 0.5) +
  stat_summary(fun.data = mean_sdl) +
  # geom_boxplot(width=0.3, alpha=0.5, outlier.shape = NA) +
  stat_summary(aes(group = 1), fun = "mean", geom = "line") +
  facet_wrap(~ var, scales = "free_y")
```
