==> analysis/12-VertPGLS2.Rmd <==
---
title: "PGLS analysis on vertebrae"
output:
  html_document:
    df_print: paged
---


```{r setup}
library(tidyverse) # for everything
library(geomorph) # for PGLS analysis
#library(emmeans) # doesn't work with geomorph
library(ggbeeswarm) # shows beeswarm points
library(ggpubr) # for effect size brackets
library(phytools) # for loading phylogenies
library(geiger) # phylogenetics - but we just use name.check
library(rstatix)
library(esvis)  # for effect sizes
library(Hmisc) # for simple points with confidence limits
library(patchwork) # for tiling figure panels
library(gt) # for making nice tables
library(here) # for locating data
```

Package details

```{r}
R.version.string
```

```{r}
citation()
```
```{r}
citation('geomorph')
```
```{r}
packageVersion('geomorph')
packageVersion('rrpp')
```

```{r}
citation('esvis')
```

```{r}
packageVersion('esvis')
```

# Load data

## Vertebral measurements

```{r}
vertdata <- read_csv(here('output', "MasterVert_Measurements_Matched.csv")) %>%
  # separate(MatchSpecies, into=c("MatchGenus", "MatchSpecies"), sep="_") %>%
  relocate(MatchSpecies, .after=Species) %>%
  rename(alphaPos = alpha_Pos,
         alphaAnt = alpha_Ant,
         DPos = D_Pos,
         DAnt = D_Ant,
         BodyShape = `Body Shape`,
         dBW = d_BW,
         DAntBW = D_Ant_BW,
         DPosBW = D_Pos_BW,
         fineness = `SL/Max_BW`)

head(vertdata)
```

Pull out the landmarks for each vertebra and scale each to body length.
```{r}
vertdatapts <-
  vertdata |> 
  filter(Indiv == 1) |> 
  select(Habitat, Species, MatchSpecies, Pos, starts_with('Pt'), fineness, SL) |>
  pivot_longer(cols = starts_with('Pt'),
               names_to = c('point', '.value'),
               names_pattern = "Pt(.)(.)") |>
  mutate(x = x/SL,
         y = y/SL)
  
```

Get the most common value
```{r}
mode <- function(x) {
   return(as.numeric(names(which.max(table(x)))))
}
```


```{r}
toofewpoints <-
  vertdatapts |> 
  filter(Pos %in% seq(40,90, by = 10) & !is.na(Habitat)) |> 
  group_by(Species) |> 
  dplyr::summarize(npts = sum(!is.na(x))) |> 
  ungroup() |> 
  filter(npts != mode(npts)) |> 
  pull(Species)
```

Get species that have the same number of vertebrae digitized.
```{r}
vertdatapts <-
  vertdatapts |> 
  filter(Pos %in% seq(40,90, by = 10) & !is.na(Habitat)) |> 
  group_by(Species) |> 
  mutate(npts = sum(!is.na(x))) |> 
  ungroup() |> 
  filter(npts == mode(npts))
  
```

## Summary data

These data tables have means for each species, plus the pruned phylogenetic tree.

```{r}
vertdata_sp <- read_csv(here('output/vertdata_summary_species.csv')) |> 
  mutate(Habitat = factor(Habitat, levels = c('benthic', 'demersal', 'pelagic')))
verttree <- readRDS(here('output/vert_tree.rds'))
```

```{r}
vertdata_sp <-
  vertdata_sp |> 
  filter(!(Species %in% toofewpoints))
```


```{r}
vertdatapts <-
  vertdatapts |> 
  filter(Species %in% vertdata_sp$Species)
```


# Tests on means

## Construct the geomorph data frame

`geomorph` requires a sort of weird version of the data frame, including multidimensional arrays with dimensions labeled according to the species name. This constructs the data frame containing the variables we're interested in.

```{r}
vars <- c('CBL_mn', 'd_mn', 'alphaAnt_mn', 'alphaPos_mn', 'DAnt_mn', 'DPos_mn',
         'fineness', 'Habitat')

gdf <- geomorph.data.frame(phy = verttree)
for (v in vars) {
  arr = array(vertdata_sp[[v]], dim = c(nrow(vertdata_sp), 1),
              dimnames = list(vertdata_sp$Species, NULL))
  gdf[[v]] <- arr
}
```


# Check fineness

How does fineness depend on habitat?

```{r}
ggplot(vertdata_sp, aes(x = Habitat, y = fineness, color = Habitat, shape = Habitat)) +
  geom_quasirandom(width=0.3, alpha = 0.5) +
  stat_summary(fun.data = 'mean_cl_boot') +
  stat_summary(aes(group = 1), fun = 'mean', geom='line', color='black') +
  labs(y = "Fineness") +
  scale_x_discrete(labels = c("b", "d", "p")) +
  scale_shape_manual(values = c(15, 19, 4)) +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  theme_bw() + theme(aspect.ratio = 0.7)
```


```{r}
fineness_habitat <- procD.pgls(fineness ~ Habitat, phy = phy, data = gdf, iter = 999, 
                               SS.type = 'III', print.progress = FALSE)
```

```{r}
summary(fineness_habitat)
```
Marginally significant, with pelagic species more elongate than benthic.


# Multivariate ANOVA

Putting all of our measurements together, is there a difference relative to habitat or fineness?

```{r}
manova.pgls <- procD.pgls(as.matrix(cbind(CBL_mn, d_mn, alphaAnt_mn, alphaPos_mn, DAnt_mn, DPos_mn))
                     ~ Habitat + fineness, phy = phy, data = gdf, iter = 999, SS.type = 'III',
                     print.progress = FALSE)
```

```{r}
manova.pgls <- RRPP::manova.update(manova.pgls, tol = 0, print.progress = FALSE)
```


```{r}
manova.sum <- summary(manova.pgls, test = 'Pillai')
manova.sum
```


# Multivariate tests on individual vertebrae



Get the names of each species, without repeats.
```{r}
ids <-
  distinct(vertdatapts, Species, .keep_all=TRUE)
```

```{r}
nrow(ids)
```

```{r}
ids <-
  ids |> 
  mutate(rowname = Species) %>%
  remove_rownames() |> 
  column_to_rownames(var = "rowname")
```


```{r}
name.check(verttree, ids)
```

```{r}
verttree2 <- keep.tip(verttree, ids$Species)
```

```{r}
vertdatapts <-
  vertdatapts |> 
  group_by(Habitat, fineness, SL, Species, MatchSpecies, Pos) |> 
  mutate(x = x - mean(x),
         y = y - mean(y))
```

## Construct geomorph data frames for each vertebra

This is the data frame with all of the vertebrae together.
```{r}
pts <- with(vertdatapts,
            array(data = c(x, y),
             dim = c(mode(npts), 2, nrow(ids)),
             dimnames = list(NULL, NULL, ids$Species)))
Y.gpa <- gpagen(pts, print.progress = FALSE)

gdfall <- geomorph.data.frame(Y.gpa, phy = verttree2)
gdfall$Habitat = array(ids$Habitat, dim = c(nrow(ids), 1),
                     dimnames = list(ids$Species, NULL))
gdfall$fineness = array(ids$fineness, dim = c(nrow(ids), 1),
                      dimnames = list(ids$Species, NULL))
```

```{r}
pgall <- procD.pgls(coords ~ Habitat + fineness, phy = verttree,
                    data = gdfall, inter = 999, SS.type = 'III',
                    print.progress = FALSE)
pgall <- RRPP::manova.update(pgall, tol = 0, print.progress = FALSE)
```


```{r}
summary(pgall, test = 'Pillai')
```

Pull out vertebrae at 40%, 50%, etc and construct the data frame for each one separately.
```{r}
gdf.vert1 <- list()
selpts <- seq(40,90, by = 10)

for (i in seq_along(selpts)) {
  # print(selpts[[i]])
  v1 <- vertdatapts |> 
    filter(Pos == selpts[[i]]) |> 
    group_by(Species) |> 
    mutate(n1 = n())
  
  if (any(v1$n1 != mode(v1$n1))) stop('Bad!') 
  
  # print(v1 |> select(Pos, x,y) |> head())
  pts <- array(data = c(v1$x, v1$y),
          dim = c(mode(v1$n1), 2, nrow(ids)),
             dimnames = list(NULL, NULL, ids$Species))
  Y.gpa1 <- gpagen(pts, print.progress = FALSE)
  # print(head(Y.gpa1$consensus))
  
  gdf1 <- geomorph.data.frame(Y.gpa1, phy = verttree2)
  gdf1$Habitat <- array(ids$Habitat, dim = c(nrow(ids), 1),
                       dimnames = list(ids$Species, NULL))
  gdf1$fineness <- array(ids$fineness, dim = c(nrow(ids), 1),
                        dimnames = list(ids$Species, NULL))
  gdf.vert1[[i]] <- gdf1
}
```

Run the multivariate PGLS tests for each vertebra.
```{r}
pgls.vert1 <- list()
for (i in seq_along(gdf.vert1)) {
  pg1 <- procD.pgls(coords ~ Habitat + fineness, phy = phy, 
                    data = gdf.vert1[[i]],
                    iter = 999, SS.type = 'III',
                    print.progress = FALSE)
  pgls.vert1[[i]] <- pg1
  pgls.vert1[[i]] <- RRPP::manova.update(pg1, tol = 0, print.progress = FALSE)
}
```

```{r}
all.summaries <- purrr::map(pgls.vert1, ~ summary(.x, test = 'Pillai'))
```

```{r}
manovatabs <-
  purrr::map2(all.summaries, selpts, ~ .x$stats.table |> 
              mutate(var = as.character(.y)) |> 
              rownames_to_column(var = 'Effect')) |> 
  bind_rows()

manovatabs
```



```{r}
manova.all.sum <- summary(pgall, test = 'Pillai')

manovatabsall <-
  manova.all.sum$stats.table |> 
  as.data.frame() |> 
  rownames_to_column(var = 'Effect') |> 
  mutate(var = 'All vertebrae') |> 
  bind_rows(manovatabs)

manovatabsall
```

```{r manova_table}
manovatab <-
  manovatabsall |> 
  rename(p = `Pr(>Pillai)`) |> 
  select(var, Effect, Df, Pillai, Z, p) |> 
  filter(Effect %in% c('Habitat', 'fineness', 'Habitat:fineness')) |> 
  group_by(var) |> 
  gt(
    groupname_col = "var",
    rowname_col = "Effect"
  ) |> 
  fmt_number(
    columns = c("Pillai", "Z"),
    suffixing = FALSE,
    n_sigfig = 2
  ) |> 
  fmt_number(
    columns = "p",
    decimals = 3
  ) |> 
  cols_label(
    var = md("Location"),
    p = md("p"),
  ) |> 
  tab_style(
    locations = cells_column_labels(columns = c("var", 'Df', 'Pillai', 'Z', "p")),
    style = cell_text(v_align = "middle",
                      align = "center")
  ) |> 
  tab_stubhead("Location") |> 
  tab_style(
    locations = cells_stubhead(),
    style = cell_text(v_align = "middle")
  ) |> 
  sub_missing(columns = 3:6,
              missing_text = "")  

manovatab |> 
  as_raw_html()
```

```{r}
gtsave(manovatab, here("output/manova_table.rtf"))
```


# Univariate tests

Test each measurement separately.

```{r}
models <- list(CBL_mn = procD.pgls(CBL_mn ~ Habitat + fineness, phy = phy, data = gdf, 
                                   iter = 999, SS.type = 'III', print.progress = FALSE),
            d_mn = procD.pgls(d_mn ~ Habitat + fineness, phy = phy, data = gdf, 
                              iter = 999, SS.type = 'III', print.progress = FALSE),
            alphaAnt_mn = procD.pgls(alphaAnt_mn ~ Habitat + fineness, phy = phy, data = gdf, 
                                     iter = 999, SS.type = 'III', print.progress = FALSE),
            alphaPos_mn = procD.pgls(alphaPos_mn ~ Habitat + fineness, phy = phy, data = gdf, 
                                     iter = 999, SS.type = 'III', print.progress = FALSE),
            DAnt_mn = procD.pgls(DAnt_mn ~ Habitat + fineness, phy = phy, data = gdf, 
                                 iter = 999, SS.type = 'III', print.progress = FALSE),
            DPos_mn = procD.pgls(DPos_mn ~ Habitat + fineness, phy = phy, data = gdf, 
                                 iter = 999, SS.type = 'III', print.progress = FALSE),
            fineness = procD.pgls(fineness ~ Habitat, phy = phy, data = gdf, 
                                 iter = 999, SS.type = 'III', print.progress = FALSE))
```

```{r}
habdf <- with(vertdata_sp,
               expand_grid(Habitat = levels(Habitat)))
finedf <- with(vertdata_sp,
               expand_grid(fineness = pracma::linspace(min(fineness), max(fineness), n = 10)))
habfinedf <- with(vertdata_sp,
               expand_grid(Habitat = levels(Habitat),
                           fineness = pracma::linspace(min(fineness), max(fineness), n = 10)))
```


```{r}
anovatabs <- c()
predvals <- tibble(var = names(models),
                   data = list(list()))
habvals <- tibble(var = names(models),
                  data = list(list()))
for (i in seq_along(models)) {
  print(names(models)[i])
  tab1 <- data.frame(models[[i]]$aov.table)
  tab1$var <- names(models)[i]
  anovatabs[[i]] <- tab1 |> 
    rownames_to_column(var = 'Effect') |> 
    rename(p = Pr..F.)
  
  # pint <- anovatabs[[i]] |> 
  #   filter(Effect == 'Habitat:fineness') |> 
  #   pull(p)
  phab <- anovatabs[[i]] |> 
    filter(Effect == 'Habitat') |> 
    pull(p)
  pfine <- anovatabs[[i]] |> 
    filter(Effect == 'fineness') |> 
    pull(p)

  if (length(pfine) == 0) {
    pfine <- Inf
  }
  pred1 <- predict(models[[i]], newdata = habdf)
  
  pred1df <- habdf
  pred1df$pred <- as.vector(pred1$mean)
  pred1df$lcl <- as.vector(pred1$lcl)
  pred1df$ucl <- as.vector(pred1$ucl)
  
  habvals$data[[i]] <- pred1df
  
  if (phab < 0.05 & pfine < 0.05) {
    print('both')
    pred1 <- predict(models[[i]], newdata = habfinedf)

    pred1df <- habfinedf
    pred1df$pred <- as.vector(pred1$mean)

    predvals$data[[i]] <- pred1df
  } else if (phab < 0.05) {
    print('habitat')
    pred1 <- predict(models[[i]], newdata = habdf)
  
    pred1df <- habdf
    pred1df$pred <- as.vector(pred1$mean)
    pred1df$lcl <- as.vector(pred1$lcl)
    pred1df$ucl <- as.vector(pred1$ucl)
    
    predvals$data[[i]] <- pred1df
  } else if (pfine < 0.05) {
    print('fineness')
    pred1 <- predict(models[[i]], newdata = finedf)
  
    pred1df <- finedf
    pred1df$pred <- as.vector(pred1$mean)
    pred1df$lcl <- as.vector(pred1$lcl)
    pred1df$ucl <- as.vector(pred1$ucl)
    predvals$data[[i]] <- pred1df
  } else {
    print('none')
    pred1df <- habdf
    pred1df$pred <- NA
    pred1df$lcl <- NA
    pred1df$ucl <- NA
    predvals$data[[i]] <- pred1df
  }
  
}

anovatabs <- bind_rows(anovatabs) 
anovatabs
```

```{r}
habvals <-
  habvals |> 
  unnest(data)

write_csv(habvals, here('output/habitatvals.csv'))
```

```{r}
saveRDS(predvals, file = here('output/predvals.Rds'))
```

## Construct effect sizes

```{r}
vars <- c('CBL_mn', 'd_mn', 'alphaAnt_mn', 'alphaPos_mn', 'DAnt_mn', 'DPos_mn',
         'fineness', 'Habitat')
pairs <- list(c('benthic','demersal'),
              c('benthic','pelagic'),
              c('demersal','pelagic'))

gdfpair <- list()
for (i in seq_along(pairs)) {
  gdfpair[[i]] <- geomorph.data.frame(phy = verttree)
  vp1 <- vertdata_sp |> 
    filter(Habitat %in% pairs[[i]])
  
  for (v in vars) {
    arr = array(vp1[[v]], dim = c(nrow(vp1), 1),
                dimnames = list(vp1$Species, NULL))
    gdfpair[[i]][[v]] <- arr
  }
}
```

```{r}
modelpairs <- list(CBL_p1 = procD.pgls(CBL_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[1]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   CBL_p2 = procD.pgls(CBL_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[2]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   CBL_p3 = procD.pgls(CBL_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[3]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   d_p1 = procD.pgls(d_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[1]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   d_p2 = procD.pgls(d_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[2]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   d_p3 = procD.pgls(d_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[3]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   alphaAnt_p1 = procD.pgls(alphaAnt_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[1]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   alphaAnt_p2 = procD.pgls(alphaAnt_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[2]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   alphaAnt_p3 = procD.pgls(alphaAnt_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[3]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   alphaPos_p1 = procD.pgls(alphaPos_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[1]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   alphaPos_p2 = procD.pgls(alphaPos_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[2]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   alphaPos_p3 = procD.pgls(alphaPos_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[3]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),                   
                   DAnt_p1 = procD.pgls(DAnt_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[1]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   DAnt_p2 = procD.pgls(DAnt_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[2]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   DAnt_p3 = procD.pgls(DAnt_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[3]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   DPos_p1 = procD.pgls(DPos_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[1]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   DPos_p2 = procD.pgls(DPos_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[2]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   DPos_p3 = procD.pgls(DPos_mn ~ Habitat + fineness, phy = phy, data = gdfpair[[3]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   fineness_p1 = procD.pgls(fineness ~ Habitat, phy = phy, data = gdfpair[[1]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   fineness_p2 = procD.pgls(fineness ~ Habitat, phy = phy, data = gdfpair[[2]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE),
                   fineness_p3 = procD.pgls(fineness ~ Habitat, phy = phy, data = gdfpair[[3]],
                                       inter = 999, SS.type = 'III', print.progress = FALSE))
                   
```

```{r}
effectsizes <- list()
for (i in seq_along(modelpairs)) {
  a1 <- anova(modelpairs[[i]], effect.type = 'cohenf')
  df1 <- as_tibble(a1$table) |> 
    rename(p = `Pr(>Cohen's f-squared)`,
           eff = Z) |> 
    select(eff, p) |> 
    head(n = 1) |> 
    mutate(var = names(modelpairs)[[i]])
  effectsizes[[i]] <- df1
}

effectsizes <-
  bind_rows(effectsizes) |> 
  separate(var, into = c('var', 'pair')) |> 
  mutate(Habitat_foc = case_when(pair == 'p1'  ~  'benthic',
                                 pair == 'p2'  ~  'benthic',
                                 pair == 'p3'  ~  'demersal'),
         Habitat_ref = case_when(pair == 'p1'  ~  'pelagic',
                                 pair == 'p2'  ~  'demersal',
                                 pair == 'p3'  ~  'pelagic'),
         Habitat = Habitat_ref) |> 
  rename(group1 = Habitat_ref,
         group2 = Habitat_foc) |> 
  select(-pair)
```

```{r}
effectsizes
```


```{r}
write_csv(effectsizes, here('output/effectsizes.csv'))
```

Rearrange the table
```{r}
effectsizes_wide <- 
  effectsizes |> 
  mutate(contrast = str_c(group1, group2, sep = ' - '),
         contrast = str_replace(contrast, '(\\w+) - (\\w+)', '\\1_\\2')) |> 
  select(var, contrast, eff, p) |> 
  pivot_wider(names_from = contrast, values_from = c(eff, p),
              names_glue = '{contrast}_{.value}') |> 
  mutate(var = if_else(var == 'fineness', var, str_c(var, '_mn')),
         Effect = 'Habitat')

effectsizes_wide
```

```{r}
anovatabs <-
  anovatabs |> 
  left_join(effectsizes_wide, by = c('var', 'Effect'))
```

```{r}
write_csv(anovatabs, here("output/anovatabs.csv"))
```


## Make the stats table

```{r maketable}
anovatab <-
  anovatabs |> 
  select(var, Effect, Df, Rsq, F, Z, p, ends_with('p')) |> 
  filter(Effect %in% c('Habitat', 'fineness', 'Habitat:fineness')) |> 
  mutate(var = str_remove(var, '_mn')) |> 
  mutate(var = case_when(
    var == "alphaAnt"   ~   "Anterior cone angle",
    var == "alphaPos"   ~   "Posterior cone angle",
    var == "CBL"   ~   "Centrum body length",
    var == "DAnt"   ~   "Anterior cone diameter",
    var == "DPos"   ~   "Posterior cone diameter",
    var == "d"   ~   "Foramen diameter",
    var == "fineness"   ~   "Fineness"
  )) |> 
  #group_by(var) |> 
  gt(
    # groupname_col = "var",
    rowname_col = "Effect"
  ) |> 
  fmt_number(
    columns = c("F", "Z"),
    suffixing = FALSE,
    n_sigfig = 2
  ) |> 
  fmt_number(
    columns = "p",
    decimals = 3
  ) |> 
  fmt_number(
    columns = "Rsq",
    decimals = 2
  ) |> 
  fmt_number( 
    columns = c('demersal_benthic_p', 'pelagic_benthic_p', 'pelagic_demersal_p'),
    # columns = c('demersal_benthic_p', 'pelagic_benthic_p', 'pelagic_demersal_p'),
    # columns = c("benthic_demersal_eff", "benthic_pelagic_eff", "demersal_pelagic_eff"),
    decimals = 3,
    # force_sign = TRUE
  ) |> 
  cols_label(
    var = md("Measurement"),
    p = md("p"),
    demersal_benthic_p = md("d - b"),
    pelagic_benthic_p = md("p - b"),
    pelagic_demersal_p = md("p - d")
  ) |> 
  tab_style(
    locations = cells_column_labels(columns = c("var", 'Df', 'Rsq', 'F', 'Z', "p")),
    style = cell_text(v_align = "middle",
                      align = "center")
  ) |> 
  #tab_stubhead("Measurement") |> 
  tab_style(
    locations = cells_stubhead(),
    style = cell_text(v_align = "middle")
  ) |> 
  # tab_spanner(
  #   label = "Effect sizes",
  #   columns = c("benthic_demersal_eff", "benthic_pelagic_eff", "demersal_pelagic_eff")
  # ) |> 
  sub_missing(columns = 2:10,
              missing_text = "")
  
anovatab |> 
  as_raw_html()
```

```{r}
gtsave(anovatab, here("output/anova_table.rtf"))
```
