==> analysis/20-plot_phylogeny.Rmd <==
---
title: "Make Figure 1"
output: html_notebook
---

```{r setup}
library(tidyverse)
library(ggbeeswarm)
library(phytools)
library(patchwork)
library(here)
library(ggtree)
library(plotly)
```

## Figure 1

For this figure, we need to identify three species from the three habitat classes that have clearly different vertebrae.

```{r}
vertdata <- read_csv(here('output/vertdata_summary_species.csv'))
```

```{r}
plot_ly(data = vertdata, type = "scatter", mode = "markers") %>%
  add_trace(x = ~Habitat, y = ~d_med, type = "box",
            text = ~Species, hoverinfo = "text",
            boxpoints = "all", jitter = 0.2)
```

```{r echo=FALSE, results='asis'}
vertdata %>%
  group_by(Habitat) %>%
  mutate(d_med_group = median(d_med, na.rm = TRUE)) %>%
  filter(abs(d_med - d_med_group)/d_med_group < 0.1) %>%
  ungroup() %>%
  arrange(Habitat)
```

Choose example species close to the median for their group:

* benthic: Barbichthys laevis (Sucker barb) or Myoxocephalus polyacanthocephalus (Sculpin)
* demersal: Poecilia reticulata (Guppy)
* pelagic: Sphyraena sphyraena (Barracuda)

We'll use the sculpin as an example benthic species, because we have good histology data for it.

```{r}
examplespecies <- list("Myoxocephalus_polyacanthocephalus",
                    "Anoplarchus_purpurescens",
                    "Cymatogaster_aggregata")
```

```{r}
verttree <- readRDS(here('output/vert_tree.rds'))
```

```{r}
vertdata %>%
  filter(Species %in% examplespecies)
```
```{r}
vertdata <-
  vertdata %>%
  mutate(WaterTypeShort = str_sub(Water_Type, start = 1, end = 1))
```

```{r}
vertdata |> 
  distinct(Species) |> 
  dplyr::summarise(n = n())
```

```{r}
vertdata |> 
  group_by(Habitat) |> 
  dplyr::summarize(n = n()) |> 
  mutate(pct = n / sum(n) * 100)
```

```{r}
vertdata |> 
  distinct(Family) |> 
  dplyr::summarise(n = n())
```

```{r}
vertdata |> 
  group_by(Habitat) |> 
  distinct(Family) |> 
  dplyr::summarize(n = n())
```

```{r}
vertdata |> 
  group_by(Water_Type) |> 
  dplyr::summarize(n = n()) |> 
  mutate(pct = n / sum(n) * 100)
```

```{r}
orders <-
  left_join(as_tibble(verttree), vertdata, by = c("label" = "Species")) |> 
  mutate(Species = label,
         label = str_replace(label, '_', ' ')) |> 
  group_by(Order) |> 
  dplyr::summarize(id = min(parent),
                   n = n()) |> 
  filter(n >= 2 & !str_detect(Order, 'Incertae') & !is.na(Order))
orders
```
```{r}
nodestolabel <- c('Actinopterygii',
                  'Neopterygii',
                  'Teleostei',
                  'Otomorpha',
                  # 'Euteleostomorpha',
                  'Neoteleostei',
                  'Acanthomorphata',
                  'Percomorphaceae',
                  'Eupercaria')

allnodes <-
  left_join(as_tibble(verttree), vertdata, by = c("label" = "Species")) |> 
  mutate(Species = label,
         label = str_replace(label, '_', ' '),
         alltaxon = replace_na(alltaxon, '-')) |> 
  select(parent, node, alltaxon)

labelnodes <- tibble()
for (n in nodestolabel) {
  print(n[[1]])
  labelnodes <-
    allnodes |>
    dplyr::filter(str_detect(alltaxon, n[[1]])) |> 
    dplyr::summarize(taxon = n[[1]],
                     # alltaxon = alltaxon[1],
                     pmin = min(parent),
                     nmin = min(node)) |> 
    bind_rows(labelnodes)
}
labelnodes
```


```{r}
left_join(as_tibble(verttree), vertdata, by = c("label" = "Species")) %>%
  # left_join(labelnodes, by = c('parent' = 'pmin')) %>%
  mutate(Species = label,
         label = str_replace(label, '_', ' ')) %>%
         # str_c(Family, Species)) %>%
  tidytree::as.treedata() %>%
  ggtree() + #branch.length = 'none') + # layout = "circular", open.angle = 120) + 
  scale_y_reverse() +
  geom_tiplab(aes(color = Habitat), size=1.5, offset = 0.2) +
  geom_tippoint(aes(shape = Water_Type)) +
  geom_text2(aes(label=label, subset=Species %in% examplespecies),
             hjust = 0, vjust = 0) +
  geom_cladelab(data=orders,
                mapping=aes(node=id, label=Order), geom='text',
                offset=50) +
  # geom_text2(aes(label=taxon, subset=!is.na(taxon))) +
  geom_treescale() +
  scale_shape_manual(values = c(3, 23, 24)) +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  theme(legend.position = "bottom")
  #geom_label2(aes(label='P', subset = ispair))
```


```{r}
ltt.coplot(verttree, show.tip.label = FALSE)
```


```{r}
# ggsave(here('output/phylogeny_families.pdf'), width=6.5, height=8, units="in")

ggsave(here('output/plot_example_data_figure.pdf'), width=3.5, height=6, units="in")
```

```{r}
vertdata %>%
  group_by(Habitat) %>%
  summarize(n = n(), frac = n() / nrow(vertdata))
```

```{r}
vertdata %>%
  group_by(Water_Type) %>%
  summarize(n = n(), frac = n() / nrow(vertdata))
```
