==> analysis/21-plot_fits_and_summary.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r setup}
library(tidyverse)
library(ggbeeswarm)
library(patchwork)
library(emmeans)
library(gt)
library(here)
```

```{r}
vertdata_sp <- read_csv(here('output/vertdata_summary_lm_species.csv')) |> 
  mutate(Habitat = factor(Habitat, levels = c('benthic', 'demersal', 'pelagic')))
vertdata_all <- read_csv(here("output/vertdata_centered.csv"))
anovatabs <- read_csv(here("output/anovatabs.csv"))
```

```{r}
reversemeasurements <- function(df) {
  df %>%
    mutate(Pt1x = CBL/2,
           Pt1y = DPos/2,
           Pt2x = CBL/2,
           Pt2y = -DPos/2,
           Pt3x = -CBL/2,
           Pt3y = DAnt/2,
           Pt4x = -CBL/2,
           Pt4y = -DAnt/2,
           Pt5x = 0,
           Pt5y = 0,
           Pt6x = 0,
           Pt6y = d/2,
           Pt7x = 0,
           Pt7y = -d/2,
           Pt8x = CBL/2 - (DPos - d)/2 / tan(alphaPos/2 * pi/180),
           Pt8y = d/2,
           Pt9x = CBL/2 - (DPos - d)/2 / tan(alphaPos/2 * pi/180),
           Pt9y = -d/2,
           Pt10x = -CBL/2 + (DAnt - d)/2 / tan(alphaAnt/2 * pi/180),
           Pt10y = d/2,
           Pt11x = -CBL/2 + (DAnt - d)/2 / tan(alphaAnt/2 * pi/180),
           Pt11y = -d/2)
}
```

```{r}
habitatvals <- read_csv(here('output/habitatvals.csv'))
```
```{r}
vertshape <-
  habitatvals |> 
  mutate(var = str_remove(var, '_mn')) |> 
  select(-lcl, -ucl) |> 
  pivot_wider(names_from = var, values_from = pred) |> 
  reversemeasurements()
```


```{r}
vertshape <-
  vertshape %>%
  mutate(Shape1x = Pt1x,
         Shape2x = Pt2x,
         Shape3x = Pt9x,
         Shape4x = Pt11x,
         Shape5x = Pt4x,
         Shape6x = Pt3x,
         Shape7x = Pt10x,
         Shape8x = Pt8x,
         Shape9x = Pt1x,
         Shape1y = Pt1y,
         Shape2y = Pt2y,
         Shape3y = Pt9y,
         Shape4y = Pt11y,
         Shape5y = Pt4y,
         Shape6y = Pt3y,
         Shape7y = Pt10y,
         Shape8y = Pt8y,
         Shape9y = Pt1y)
```

```{r}
vertshape <-
  vertshape %>%
  select(!starts_with("Pt")) %>%
  pivot_longer(starts_with("Shape"), names_to = "Pt", values_to = "value") %>%
  extract(Pt, into = c("Num", "XY"), regex = "Shape(\\d+)(x|y)") %>%
  pivot_wider(names_from = XY, values_from = value)

```

```{r}
vertshape_panel <-
  vertshape %>%
  ggplot(aes(x = x, y = y, color = Habitat, fill = Habitat)) +
  geom_path() +
  geom_polygon(alpha = 0.2) +
  coord_fixed() +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2"),
                     guide = "none") +
  scale_fill_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2"),
                     guide = "none") +
  annotate("line", x = c(0.01, 0.01), y = c(-0.003, -0.007), color = "black") +
  annotate("text", x = 0.01, y = -0.005, label = "0.004 BL", angle=90, size=4, vjust = 1.3) +
  theme_minimal() +
  theme(line = element_blank(),
        text = element_blank(),
        title = element_blank())

vertshape_panel
```

```{r}
ggsave(here('output/vertshape.pdf'), plot = vertshape_panel,
       width = 3, height = 2, units = "in")
```

```{r}
effectsizes <- read_csv(here('output/effectsizes.csv'))
```
```{r}
predvals <- readRDS(here('output/predvals.Rds'))
```

```{r}
dyl = 0.0005
yl = max(vertdata_sp$fineness) + dyl

pred1 <- predvals |> 
  filter(var == 'fineness') |> 
  pull(data)
pred1 <- pred1[[1]]

fineness_panel <- ggplot(vertdata_sp, aes(x = Habitat, y = fineness, color = Habitat, shape = Habitat)) +
  geom_quasirandom(width=0.3, alpha = 0.5) +
  # stat_summary(aes(group = 1), fun = mean, geom = 'line', color = 'black') +
  # stat_summary(fun.data = 'mean_cl_boot', color = 'black') +
  stat_pvalue_manual(data = filter(effectsizes, var == 'fineness',
                                   p < 0.05),
                     y.position = yl, step.increase = 0.12,
                     label = '{round(p, digits=3)}', size = 3) +
  geom_point(data = pred1, inherit.aes = FALSE, aes(x = Habitat, y = pred, ymin = lcl, ymax = ucl), color = 'black') +
  geom_line(data = pred1, inherit.aes = FALSE, aes(x = Habitat, y = pred, group = 1), color = 'black') +
  labs(y = "Fineness ratio") +
  scale_x_discrete(labels = c("b", "d", "p")) +
  scale_shape_manual(values = c(15, 19, 4), guide='none') +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2"),
                     guide='none') +
  theme_bw() + theme(aspect.ratio = 0.7,
                     legend.position = "bottom",
                     panel.border = element_blank(), axis.line = element_line())


fineness_panel
```

```{r}
ggsave(here('output/fineness.pdf'), plot = fineness_panel,
       width = 3, height = 2, units = "in")
```

```{r}
dyl = 0.0005
yl = max(vertdata_sp$d_mn) + dyl

phab <- anovatabs |> 
  filter(var == 'd_mn' & Effect == 'Habitat') |> 
  pull(p)
pfine <- anovatabs |> 
  filter(var == 'd_mn' & Effect == 'fineness') |> 
  pull(p)
pred1 <- predvals |> 
  filter(var == 'd_mn') |> 
  pull(data)
pred1 <- pred1[[1]]

d_mn_panel <- ggplot(vertdata_sp, aes(x = Habitat, y = d_mn, color = Habitat, shape = Habitat)) +
  geom_quasirandom(width=0.3, alpha = 0.5) +
  # stat_summary(aes(group = 1), fun = mean, geom = 'line', color = 'black') +
  # stat_summary(fun.data = 'mean_cl_boot', color = 'black') +
  stat_pvalue_manual(data = filter(effectsizes, var == 'd',
                                   p < 0.05),
                     y.position = yl, step.increase = 0.12,
                     label = '{round(p, digits=3)}', size = 3) +
  geom_point(data = pred1, inherit.aes = FALSE, aes(x = Habitat, y = pred, ymin = lcl, ymax = ucl), color = 'black') +
  geom_line(data = pred1, inherit.aes = FALSE, aes(x = Habitat, y = pred, group = 1), color = 'black') +
  annotate('text', x = -Inf, y = Inf, hjust = -0.1, vjust = 1.5, 
           label = glue::glue('p_fine = {pfine}; p_hab = {phab}'), size = 3) +
  labs(y = "Mean foramen\ndiameter (BL)") +
  scale_x_discrete(labels = c("b", "d", "p")) +
  scale_shape_manual(values = c(15, 19, 4)) +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  theme_bw() + theme(aspect.ratio = 0.7)

d_mn_panel
```
```{r}
dyl = 0.0005
yl = max(vertdata_sp$d_mn) + dyl

phab <- anovatabs |> 
  filter(var == 'alphaAnt_mn' & Effect == 'Habitat') |> 
  pull(p)
pfine <- anovatabs |> 
  filter(var == 'alphaAnt_mn' & Effect == 'fineness') |> 
  pull(p)
pred1 <- predvals |> 
  filter(var == 'alphaAnt_mn') |> 
  pull(data)
pred1 <- pred1[[1]]

alphaAnt_mn_panel <- ggplot(vertdata_sp, aes(x = Habitat, y = alphaAnt_mn, color = Habitat, shape = Habitat)) +
  geom_quasirandom(width=0.3, alpha = 0.5) +
  # stat_summary(aes(group = 1), fun = mean, geom = 'line', color = 'black') +
  # stat_summary(fun.data = 'mean_cl_boot', color = 'black') +
  # stat_pvalue_manual(data = filter(effectsizes, var == 'd',
  #                                  abs(eff) > 0.3),
  #                    y.position = yl, step.increase = 0.12,
  #                    label = '{round(eff, digits=2)}', size = 3) +
  # geom_point(data = pred1, inherit.aes = FALSE, aes(x = Habitat, y = pred, ymin = lcl, ymax = ucl), color = 'black') +
  # geom_line(data = pred1, inherit.aes = FALSE, aes(x = Habitat, y = pred, group = 1), color = 'black') +
  annotate('text', x = -Inf, y = Inf, hjust = -0.1, vjust = 1.5, 
           label = glue::glue('p_fine = {pfine}; p_hab = {phab}'), size = 3) +
  labs(y = "Anterior cone\nangle (deg)") +
  scale_x_discrete(labels = c("b", "d", "p")) +
  scale_shape_manual(values = c(15, 19, 4)) +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  theme_bw() + theme(aspect.ratio = 0.7)

alphaAnt_mn_panel
```

```{r}
dyl = 5
yl = max(vertdata_sp$alphaPos_mn) + dyl

phab <- anovatabs |> 
  filter(var == 'alphaPos_mn' & Effect == 'Habitat') |> 
  pull(p)
pfine <- anovatabs |> 
  filter(var == 'alphaPos_mn' & Effect == 'fineness') |> 
  pull(p)
pred1 <- predvals |> 
  filter(var == 'alphaPos_mn') |> 
  pull(data)
pred1 <- pred1[[1]]

alphaPos_mn_panel <- ggplot(vertdata_sp, aes(x = Habitat, y = alphaPos_mn, color = Habitat, shape = Habitat)) +
  geom_quasirandom(width=0.3, alpha = 0.5) +
  # stat_summary(aes(group = 1), fun = mean, geom = 'line', color = 'black') +
  # stat_summary(fun.data = 'mean_cl_boot', color = 'black') +
  stat_pvalue_manual(data = filter(effectsizes, var == 'alphaPos',
                                   p < 0.05),
                     y.position = yl, step.increase = 0.12,
                     label = '{round(p, digits=3)}', size = 3) +
  geom_point(data = pred1, inherit.aes = FALSE, aes(x = Habitat, y = pred, ymin = lcl, ymax = ucl), color = 'black') +
  geom_line(data = pred1, inherit.aes = FALSE, aes(x = Habitat, y = pred, group = 1), color = 'black') +
  annotate('text', x = -Inf, y = Inf, hjust = -0.1, vjust = 1.5, 
           label = glue::glue('p_fine = {pfine}; p_hab = {phab}'), size = 3) +
  labs(y = "Posterior cone\nangle (deg)") +
  scale_x_discrete(labels = c("b", "d", "p")) +
  scale_shape_manual(values = c(15, 19, 4)) +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  theme_bw() + theme(aspect.ratio = 0.7)

alphaPos_mn_panel
```

## Effects relative to fineness

```{r}
phab <- anovatabs |> 
  filter(var == 'CBL_mn' & Effect == 'Habitat') |> 
  pull(p)
pfine <- anovatabs |> 
  filter(var == 'CBL_mn' & Effect == 'fineness') |> 
  pull(p)
pred1 <- predvals |> 
  filter(var == 'CBL_mn') |> 
  pull(data)
pred1 <- pred1[[1]]

CBL_mn_panel <- ggplot(vertdata_sp, aes(x = fineness, y = CBL_mn, color = Habitat, shape = Habitat)) +
  # geom_smooth(aes(group = 1), color = 'black', method = 'lm') +
  geom_point() +
  #geom_point(data = pred1, inherit.aes = FALSE, aes(x = Habitat, y = pred, ymin = lcl, ymax = ucl), color = 'black') +
  geom_line(data = pred1, inherit.aes = FALSE, aes(x = fineness, y = pred, color = Habitat)) +
  annotate('text', x = -Inf, y = Inf, hjust = -0.1, vjust = 1.5, label = glue::glue('p_fine = {pfine}; p_hab = {phab}'), size = 3) +
  labs(y = "Centrum body\nlength (BL)") +
  scale_shape_manual(values = c(15, 19, 4)) +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  theme_bw() + theme(aspect.ratio = 0.7)

CBL_mn_panel
```


```{r}
ph <- anovatabs |> 
  filter(var == 'DPos_mn' & Effect == 'Habitat') |> 
  pull(p)
pf <- anovatabs |> 
  filter(var == 'DPos_mn' & Effect == 'fineness') |> 
  pull(p)
pred1 <- predvals |> 
  filter(var == 'DPos_mn') |> 
  pull(data)
pred1 <- pred1[[1]]

DPos_mn_panel <- ggplot(vertdata_sp, aes(x = fineness, y = DPos_mn, color = Habitat, fill = Habitat, shape = Habitat)) +
  # geom_smooth(method = 'lm') + #aes(group = 1), color = 'black', method = 'lm') +
  geom_point() +
  geom_line(data = pred1, inherit.aes = FALSE, aes(x = fineness, y = pred, color = Habitat)) +
  annotate('text', x = -Inf, y = Inf, hjust = -0.1, vjust = 1.5, 
           label = glue::glue('p_hab = {ph}, p_fine = {pf}'), size = 3) +
  labs(y = "Posterior cone\ndiameter (BL)") +
  scale_shape_manual(values = c(15, 19, 4)) +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  scale_fill_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  theme_bw() + theme(aspect.ratio = 0.7)

DPos_mn_panel
```

```{r}
ph <- anovatabs |> 
  filter(var == 'DAnt_mn' & Effect == 'Habitat') |> 
  pull(p)
pf <- anovatabs |> 
  filter(var == 'DAnt_mn' & Effect == 'fineness') |> 
  pull(p)
phf <- anovatabs |> 
  filter(var == 'DAnt_mn' & Effect == 'Habitat:fineness') |> 
  pull(p)
pred1 <- predvals |> 
  filter(var == 'DAnt_mn') |> 
  pull(data)
pred1 <- pred1[[1]]

DAnt_mn_panel <- ggplot(vertdata_sp, aes(x = fineness, y = DAnt_mn, color = Habitat, fill = Habitat, shape = Habitat)) +
  geom_point() +
  geom_line(data = pred1, inherit.aes = FALSE, aes(x = fineness, y = pred, color = Habitat)) +
  annotate('text', x = -Inf, y = Inf, hjust = -0.1, vjust = 1.5, 
           label = glue::glue('p_hab = {ph}, p_fine = {pf}, p_hxf = {phf}'), size = 3) +
  labs(y = "Anterior cone\ndiameter (BL)") +
  scale_shape_manual(values = c(15, 19, 4)) +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  scale_fill_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  theme_bw() + theme(aspect.ratio = 0.7)

DAnt_mn_panel
```


```{r}
((d_mn_panel / alphaAnt_mn_panel / alphaPos_mn_panel) | 
   (CBL_mn_panel / DAnt_mn_panel / DPos_mn_panel)) + 
  plot_annotation(tag_levels = 'A') +
  plot_layout(guides = 'collect') &
  theme(legend.position = "bottom",
        panel.border = element_blank(), axis.line = element_line())
```

```{r}
ggsave(here('output/mean_params.pdf'), width = 6, height = 7, units = "in")
```



# Distribution along the body

## Posterior cone angle

```{r}
alphaPosvPosPanel <-
  vertdata_all %>%
  filter(Pos >= 0.4) %>%
  ggplot(aes(x = Pos, y = alphaPos, color=Habitat, fill=Habitat, group=Habitat)) +
    stat_summary(fun.data = "mean_se", geom="ribbon", alpha=0.5) +
    stat_summary(fun = "mean", geom="line") +
  scale_shape_manual(values = c(15, 19, 4)) +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  scale_fill_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  labs(x = "Position (BL)", y = "Posterior cone\nangle (deg)") +
  theme_bw() + theme(aspect.ratio = 0.7)

alphaPosvPosPanel
```

## Centrum body length

```{r}
modeltests %>%
  filter(str_detect(var, "CBL")) %>%
  arrange(var)
```


```{r}
CBLvPosPanel <-
  vertdata_all %>%
  filter(Pos >= 0.4) %>%
  ggplot(aes(x = Pos, y = CBL, color=Habitat, fill=Habitat, group=Habitat)) +
    stat_summary(fun.data = "mean_se", geom="ribbon", alpha=0.5) +
    stat_summary(fun = "mean", geom="line") +
  scale_shape_manual(values = c(15, 19, 4)) +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  scale_fill_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  labs(x = "Position (BL)", y = "Centrum body\nlength (BL)") +
  theme_bw() + theme(aspect.ratio = 0.7)

CBLvPosPanel
```
```{r}
dyl = 1
yl = max(vertdata_sp$CBL_vtx) + dyl

CBLvtxPanel <- ggplot(vertdata_sp, aes(x = Habitat, y = CBL_vtx, color = Habitat, shape = Habitat)) +
  geom_quasirandom(width=0.3, alpha = 0.5) +
  geom_line(data = modelmeans, aes(x = Habitat, y = CBL_vtx_mn, group = 1), color = "black") +
  geom_pointrange(data = modelmeans, 
                  aes(x = Habitat, y = CBL_vtx_mn, 
                      ymin = CBL_vtx_mn - CBL_vtx_se, ymax = CBL_vtx_mn + CBL_vtx_se), size = 0.8) +
  #annotate("line", x = c(1,2), y = c(yl, yl), color = "black") +
  #annotate("text", x = 1.5, y = yl, label = "*** 0.34", size = 3, vjust = -0.2) +
  annotate("line", x = c(2,3), y = c(yl + dyl, yl + dyl), color = "black") +
  annotate("text", x = 2.5, y = yl+dyl, label = "* 0.36", size = 3, vjust = -0.2) +
  labs(y = "Fit location of\nlongest vertebra (BL)") +
  scale_x_discrete(labels = c("b", "d", "p")) +
  scale_shape_manual(values = c(15, 19, 4)) +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  theme_bw() # + theme(aspect.ratio = 0.7)

CBLvtxPanel
```

```{r}
dyl = 0.01
yl = max(vertdata_sp$CBL_quad) + dyl

CBLquadPanel <- ggplot(vertdata_sp, aes(x = Habitat, y = CBL_quad, color = Habitat, shape = Habitat)) +
  geom_quasirandom(width=0.3, alpha = 0.5) +
  geom_line(data = modelmeans, aes(x = Habitat, y = CBL_quad_mn, group = 1), color = "black") +
  geom_pointrange(data = modelmeans, 
                  aes(x = Habitat, y = CBL_quad_mn, 
                      ymin = CBL_quad_mn - CBL_quad_se, ymax = CBL_quad_mn + CBL_quad_se), size = 0.8) +
  annotate("line", x = c(2,3), y = c(yl, yl), color = "black") +
  annotate("text", x = 2.5, y = yl, label = "** -0.48", size = 3, vjust = -0.2) +
  annotate("line", x = c(1,3), y = c(yl + dyl, yl + dyl), color = "black") +
  annotate("text", x = 2, y = yl+dyl, label = "* -0.40", size = 3, vjust = -0.2) +
  labs(y = "Centrum body length\nquad. coeff. (BL^2)") +
  scale_x_discrete(labels = c("b", "d", "p")) +
  scale_shape_manual(values = c(15, 19, 4)) +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  theme_bw() # + theme(aspect.ratio = 0.7)

CBLquadPanel
```

# Foramen diameter

```{r}
dvPosPanel <-
  vertdata_all %>%
  filter(Pos >= 0.4) %>%
  ggplot(aes(x = Pos, y = d, color=Habitat, fill=Habitat, group=Habitat)) +
    stat_summary(fun.data = "mean_se", geom="ribbon", alpha=0.5) +
    stat_summary(fun = "mean", geom="line") +
  scale_shape_manual(values = c(15, 19, 4)) +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  scale_fill_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  labs(x = "Position (BL)", y = "Foramen\ndiameter (BL)") +
  theme_bw() + theme(aspect.ratio = 0.7)

dvPosPanel
```

```{r}
modeltests %>%
  filter(str_detect(var, "d_")) %>%
  arrange(var)
```

```{r}
dyl = 1
yl = max(vertdata_sp$d_vtx) + dyl

dvtxPanel <- ggplot(vertdata_sp, aes(x = Habitat, y = d_vtx, color = Habitat, shape = Habitat)) +
  geom_quasirandom(width=0.3, alpha = 0.5) +
  geom_line(data = modelmeans, aes(x = Habitat, y = d_vtx_mn, group = 1), color = "black") +
  geom_pointrange(data = modelmeans, 
                  aes(x = Habitat, y = d_vtx_mn, 
                      ymin = d_vtx_mn - d_vtx_se, ymax = d_vtx_mn + d_vtx_se), size = 0.8) +
  annotate("line", x = c(1,2), y = c(yl, yl), color = "black") +
  annotate("text", x = 1.5, y = yl, label = "*** -0.40", size = 3, vjust = -0.2) +
  annotate("line", x = c(1,3), y = c(yl + dyl, yl + dyl), color = "black") +
  annotate("text", x = 2, y = yl+dyl, label = "* -0.40", size = 3, vjust = -0.2) +
  labs(y = "Fit location of\nlargest foramen (BL)") +
  scale_x_discrete(labels = c("b", "d", "p")) +
  scale_shape_manual(values = c(15, 19, 4)) +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  theme_bw() # + theme(aspect.ratio = 0.7)

dvtxPanel
```

```{r}
dyl = 0.01
yl = max(vertdata_sp$d_quad) + dyl

dQuadPanel <- ggplot(vertdata_sp, aes(x = Habitat, y = d_quad, color = Habitat, shape = Habitat)) +
  geom_quasirandom(width=0.3, alpha = 0.5) +
  geom_line(data = modelmeans, aes(x = Habitat, y = d_quad_mn, group = 1), color = "black") +
  geom_pointrange(data = modelmeans, 
                  aes(x = Habitat, y = d_quad_mn, 
                      ymin = d_quad_mn - d_quad_se, ymax = d_quad_mn + d_quad_se), size = 0.8) +
  #annotate("line", x = c(2,3), y = c(yl, yl), color = "black") +
  #annotate("text", x = 2.5, y = yl, label = "** -0.48", size = 3, vjust = -0.2) +
  #annotate("line", x = c(1,3), y = c(yl + dyl, yl + dyl), color = "black") +
  #annotate("text", x = 2, y = yl+dyl, label = "* -0.40", size = 3, vjust = -0.2) +
  labs(y = "Foramen diameter\nquad. coeff. (BL^2)") +
  scale_x_discrete(labels = c("b", "d", "p")) +
  scale_shape_manual(values = c(15, 19, 4)) +
  scale_color_manual(values = c(benthic="chocolate4", demersal = "gold", pelagic = "deepskyblue2")) +
  theme_bw() # + theme(aspect.ratio = 0.7)

dQuadPanel
```

## Combined figure

```{r}
alphaPosvPosPanel / dvPosPanel / CBLvPosPanel + plot_annotation(tag_levels = 'A') +
  plot_layout(guides = 'collect') &
  theme(legend.position = "bottom",
        panel.border = element_blank(), axis.line = element_line())
```
```{r}
ggsave(here('output/BodyDistribution.pdf'), width = 3, units = "in")
```

```{r}
PGLSmodels
```


```{r}
tab <-
  PGLSmodels %>%
  select(var, statistic, p.value, ends_with("eff"), -total_eff) %>%
  separate(var, into = c("var", "posstat")) %>%
  mutate(posstat = if_else(var == "fineness", "mn", posstat)) %>%
  # group_by(var) %>%
  # group_modify(~ add_row(.x,.before=0)) %>%
  # mutate(posstat = if_else(is.na(posstat), "x", posstat)) %>%
  mutate(posstat = factor(posstat, levels = c("x", "mn", "vtx", "quad")),
         posstat = fct_recode(posstat, x = "x", mean = "mn", vertex = "vtx", "quad. coef." = "quad")) %>%
  ungroup() %>%
  mutate(var = case_when(
    var == "alphaAnt"   ~   "Anterior cone angle",
    var == "alphaPos"   ~   "Posterior cone angle",
    var == "CBL"   ~   "Centrum body length",
    var == "DAnt"   ~   "Anterior cone diameter",
    var == "DPos"   ~   "Posterior cone diameter",
    var == "d"   ~   "Foramen diameter",
    var == "fineness"   ~   "Fineness"
  )) %>%
  arrange(var, posstat) %>%
  gt(
    groupname_col = "var",
    rowname_col = "posstat"
  ) %>%
  fmt_number(
    columns = "statistic",
    suffixing = FALSE,
    n_sigfig = 2
  ) %>%
  fmt_number(
    columns = "p.value",
    decimals = 3
  ) %>%
  fmt_number(
    columns = c("benthic_demersal_eff", "benthic_pelagic_eff", "demersal_pelagic_eff"),
    decimals = 2,
    force_sign = TRUE
  ) %>%
  tab_style(
    style = cell_text(weight = "bold"),
    locations = cells_body(rows = p.value < 0.05)
  ) %>%
  cols_label(
    var = md("Measurement"),
    posstat = "",
    statistic = md("F2,74"),
    p.value = md("p"),
    benthic_demersal_eff = md("b - d"),
    benthic_pelagic_eff = md("b - p"),
    demersal_pelagic_eff = md("d - p")
  ) %>%
  tab_style(
    locations = cells_column_labels(columns = c("var", "posstat", "statistic", "p.value")),
    style = cell_text(v_align = "middle",
                      align = "center")
  ) %>%
  tab_stubhead("Measurement") %>%
  tab_style(
    locations = cells_stubhead(),
    style = cell_text(v_align = "middle")
  ) %>%
  tab_spanner(
    label = "Effect sizes",
    columns = c("benthic_demersal_eff", "benthic_pelagic_eff", "demersal_pelagic_eff")
  ) %>%
  sub_missing(columns = 1:7,
              missing_text = "")

tab
```

```{r}
gtsave(tab, here("output/stats_table.rtf"))
```
